1. Data type changed from int to double, I have to make sure it is correct in the next commit. 2. Mostly changes are in Renewal side.

+++ app/src/main/java/com/ab/hicarerun/network/models/WalletModel/WalletData.kt
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName
 data class WalletData (
     @SerializedName("earnedList") val earnedList : List<EarnedList>?,
     @SerializedName("burnedList") val burnedList : List<BurnedList>?,
-    @SerializedName("pointsInWallet") val pointsInWallet : Int?,
-    @SerializedName("totalReedemablePercentageInWallet") val totalReedemablePercentageInWallet : Int?,
-    @SerializedName("totalRedeemablePointsInWallet") val totalRedeemablePointsInWallet : Int?,
+    @SerializedName("pointsInWallet") val pointsInWallet : Double?,
+    @SerializedName("totalReedemablePercentageInWallet") val totalReedemablePercentageInWallet : Double?,
+    @SerializedName("totalRedeemablePointsInWallet") val totalRedeemablePointsInWallet : Double?,
     @SerializedName("pointsEarned") val pointsEarned : Int?,
     @SerializedName("pointsBurned") val pointsBurned : Int?,
     @SerializedName("historyList") val historyList : List<HistoryList>?
